refactor(race): race 내 경주 결과 출력 및 우승자 판단 로직 RaceManager로 이동
